Move test-utils into own folder
